+++ manager/container.go
@@ -42,6 +42,7 @@ import (
 )
 
 // Housekeeping interval.
+var enableLoadReader = flag.Bool("enable_load_reader", false, "Whether to enable cpu load reader")
 var HousekeepingInterval = flag.Duration("housekeeping_interval", 1*time.Second, "Interval between container housekeepings")
 
 var cgroupPathRegExp = regexp.MustCompile(`devices[^:]*:(.*?)[,;$]`)
@@ -67,7 +68,7 @@ type containerData struct {
 	lastErrorTime            time.Time
 
 	// Channel for subcontainers event
-	eventChannel chan container.SubcontainerEvent
+	eventChannelHyper chan container.SubcontainerEvent
 
 	// Decay value used for load average smoothing. Interval length of 10 seconds is used.
 	loadDecay float64
@@ -97,7 +98,7 @@ func (c *containerData) Start() error {
 	go c.housekeeping()
 
 	if c.info.Namespace != "raw" {
-		c.handler.WatchSubcontainers(c.eventChannel)
+		c.handler.WatchSubcontainers(c.eventChannelHyper)
 	}
 
 	return nil
@@ -311,16 +312,7 @@ func (c *containerData) GetProcessList(cadvisorContainer string, inHostNamespace
 	return processes, nil
 }
 
-func newContainerData(
-	containerName string,
-	memoryCache *memory.InMemoryCache,
-	handler container.ContainerHandler,
-	loadReader cpuload.CpuLoadReader,
-	logUsage bool,
-	collectorManager collector.CollectorManager,
-	maxHousekeepingInterval time.Duration,
-	allowDynamicHousekeeping bool,
-	eventChannel chan container.SubcontainerEvent) (*containerData, error) {
+func newContainerData(containerName string, memoryCache *memory.InMemoryCache, handler container.ContainerHandler, logUsage bool, collectorManager collector.CollectorManager, maxHousekeepingInterval time.Duration, allowDynamicHousekeeping bool, eventChannel chan container.SubcontainerEvent) (*containerData, error) {
 	if memoryCache == nil {
 		return nil, fmt.Errorf("nil memory storage")
 	}
@@ -338,9 +330,8 @@ func newContainerData(
 		housekeepingInterval:     *HousekeepingInterval,
 		maxHousekeepingInterval:  maxHousekeepingInterval,
 		allowDynamicHousekeeping: allowDynamicHousekeeping,
-		loadReader:               loadReader,
 		logUsage:                 logUsage,
-		eventChannel:             eventChannel,
+		eventChannelHyper:        eventChannel,
 		loadAvg:                  -1.0, // negative value indicates uninitialized.
 		stop:                     make(chan bool, 1),
 		collectorManager:         collectorManager,
@@ -348,6 +339,17 @@ func newContainerData(
 	cont.info.ContainerReference = ref
 
 	cont.loadDecay = math.Exp(float64(-cont.housekeepingInterval.Seconds() / 10))
+
+	if *enableLoadReader {
+		// Create cpu load reader.
+		loadReader, err := cpuload.New()
+		if err != nil {
+			// TODO(rjnagal): Promote to warning once we support cpu load inside namespaces.
+			glog.Infof("Could not initialize cpu load reader for %q: %s", ref.Name, err)
+		} else {
+			cont.loadReader = loadReader
+		}
+	}
 
 	err = cont.updateSpec()
 	if err != nil {
@@ -393,6 +395,16 @@ func (self *containerData) nextHousekeeping(lastHousekeeping time.Time) time.Tim
 func (c *containerData) housekeeping() {
 	// Start any background goroutines - must be cleaned up in c.handler.Cleanup().
 	c.handler.Start()
+	defer c.handler.Cleanup()
+
+	// Initialize cpuload reader - must be cleaned up in c.loadReader.Stop()
+	if c.loadReader != nil {
+		err := c.loadReader.Start()
+		if err != nil {
+			glog.Warningf("Could not start cpu load stat collector for %q: %s", c.info.Name, err)
+		}
+		defer c.loadReader.Stop()
+	}
 
 	// Long housekeeping is either 100ms or half of the housekeeping interval.
 	longHousekeeping := 100 * time.Millisecond
@@ -406,8 +418,6 @@ func (c *containerData) housekeeping() {
 	for {
 		select {
 		case <-c.stop:
-			// Cleanup container resources before stopping housekeeping.
-			c.handler.Cleanup()
 			// Stop housekeeping when signaled.
 			return
 		default: